Python script for analysis of LIS results
Reads JSON-formatted LIS data files and prints results of analysis to
stdout.
Currently analysis for PSA is built-in, but program logic may be
offloaded to smaller modules in the future. The checking mechanism for
file adherance to the schema is crude and does not cover all cases.
Known to run on Python 3.4.