Simple GUI for lisanalyze
Simple tkinter-based GUI wrapper around lisanalyze.py.
Calls lisanalyze.py using the subprocess module. Currently the location
of lisanalyze.py is assumed to be in the current working directory and
is hardcoded.
Known to run on Python 3.4 on both Linux and Windows; users must have
the Python 3 interpreter installed and the interpreter placed in $PATH.
Requires the Tk components for Python 3 (comes with the Windows version
of Python).